+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.15.0] - 2024-10-23
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `@miblanchard/react-native-slider` library.
+- [Marek Kurańda](https://github.com/mjkuranda): Rating dishes by logged-in users.
+- [Marek Kurańda](https://github.com/mjkuranda): `InputRange` component.
+
 ## [1.14.0] - 2024-10-23
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Posting (for logged-in users) and rendering comments underneath any dish.